+++ src/validators/commands/token/GetAllowanceCommandValidator.js
@@ -6,7 +6,6 @@ const getAllowanceSchema = customJoiValidator.object()
     owner: customJoiValidator.address().ethereum().required(),
     spender: customJoiValidator.address().ethereum().required(),
     token: customJoiValidator.address().ethereum().required(),
-    privateKeyFilePath: customJoiValidator.path().existFile(),
     draft: customJoiValidator.boolean(),
   })